Featherlook 1.3
Featherlook is a MS Word document search engine. The user enters a search word. Featherlook prompts the user to select a file in the directory it wishes to search (file selection is so the user can see all the files in the directory). Once directory is selected Featherlook outputs a window with file names and the sentence for each instance the search word is found in the directory/Word doc.
This version has custom windows (tkinter defaults removed overrideredirect(true)). The last window (search output) is not able to using the widgets/class that has the text boxes with file names and search word instances. I would be glad to accept any help.